telegram-bot/requirements.txt
python-telegram-bot>=21.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
anthropic>=0.42.0
